Move movie interface for movie entity to domain layer.
