+++ data/src/main/kotlin/dvachmovie/db/data/MovieEntity.kt
@@ -36,14 +36,3 @@ data class MovieEntity(@PrimaryKey @ColumnInfo(name = "movieUrl") override val m
         return result
     }
 }
-
-interface Movie {
-    val movieUrl: String
-    val previewUrl: String
-    val board: String
-    var isPlayed: Boolean
-    var date: LocalDateTime
-    val md5: String
-    val post: Long
-    val thread: Long
-}
