+++ domain/src/main/kotlin/dvachmovie/db/data/Movie.kt
@@ -1,0 +1,14 @@
+package dvachmovie.db.data
+
+import org.joda.time.LocalDateTime
+
+interface Movie {
+    val movieUrl: String
+    val previewUrl: String
+    val board: String
+    var isPlayed: Boolean
+    var date: LocalDateTime
+    val md5: String
+    val post: Long
+    val thread: Long
+}
